pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cog-replicate-logo-generator"
version = "1.0.0"
description = ""
requires-python = ">=3.8"
dependencies = ["cog", "pillow==9.5.0", "cairosvg==2.7.0"]
optional-dependencies = { dev = [
    "pylint",
    "pyright",
    "pytest",
    "pytest-asyncio",
    "ruff>=0.1.3",
] }


[tool.pytest.ini_options]
testpaths = "tests/"
